Updated/refactored code for AsyncIO utils.
